requirements.txt
asgiref==3.5.2
celery==5.2.6
cffi==1.15.1
cryptography==38.0.1
Django~=4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
dnspython==2.2.1
email-validator==1.3.0
factory-boy==3.2.1
Faker==15.1.1
idna==3.4
kombu==5.2.4
psycopg2==2.9.4
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
pika==1.3.1
pika-stubs==0.1.3
